feat: Add Bookshelf component
Add a new component to represent a Bookshelf in MyReads App. Each
Bookshelf consists of a list of Book components to be displayed
and a title for the Bookshelf. In MyReads app, we have three
instances of Bookshelf components. They hold 'Currently Reading',
'Want to Read' and 'Read' book shelves