refactor: Merge books objects into a single array
Previously, there were three static JavaScript arrays where each
array maintains a list of books and represents one of the book
shelves. This change merges all of these list into a single
list called "books" and adds a field to book objects that specifies
the book shelf on which the book is placed.