feat: Add BookshelfChanger component
Every book has a BookshelfChanger component which represents a
dropdown list in the UI. This change also handles the onChange
event for this component. When the selected value of the component
changes, the state of the component (the current selected value)
and the state of the App (the shelf of the corresponding book
in the books array) are also updated.